renamed namespace E2ETelemetryItemValidation
