+++ sdk/monitor/Azure.Monitor.OpenTelemetry.Exporter/tests/Azure.Monitor.OpenTelemetry.Exporter.Tests/E2ETelemetryItemValidation/LogsTests.cs
@@ -12,7 +12,7 @@ using Microsoft.Extensions.Logging;
 using OpenTelemetry.Logs;
 using Xunit;
 
-namespace Azure.Monitor.OpenTelemetry.Exporter.Tests.TelemetryItemValidation
+namespace Azure.Monitor.OpenTelemetry.Exporter.Tests.E2ETelemetryItemValidation
 {
     /// <summary>
     /// The purpose of these tests is to validate the <see cref="TelemetryItem"/> that is created
